fix(TRDD-GYNBICF9): make lint green — apply clippy's 53 fixes, no new allow attributes
`cargo clippy --workspace -- -D warnings` (the Makefile's lint target)
exited 101 with 53 style-lint errors at 3ecb2af, so the gate gated
nothing. 37 files were rewritten by `cargo clippy --fix`, the rest by
hand with the minimal rewrite each lint asks for. Every sort_by ->
sort_by_key rewrite keys on a Copy numeric field (stable sort, same
tie order); every question_mark rewrite was checked for a side effect
before the early return (none). Two lints exposed real defects: an
orphaned doc block for a deleted count_lines_of_code stub in
quality/debt.rs, deleted, and merge_specs's doc comment sitting above
a section header in contracts/specs.rs, moved to the fn. clippy's own
manual_find suggestion in patterns/coupling.rs does not compile
(cursor dropped while borrowed); bound to a local instead. A dead
`depth` parameter on the private collect_function_nodes was removed
with both call sites.

Verified: clippy exit 0, 0 errors; make test exit 0, 4830 + 1438
passed; allow(clippy count 9 before and after. cargo fmt --check was
already red on 188 files before this change and is not touched here
(TRDD-U5KJ5A8R); 31 of these files now carry fmt drift too.

+++ crates/tldr-cli/src/commands/remaining/todo.rs
@@ -493,7 +493,7 @@ fn run_dead_analysis(path: &Path, language: Language) -> RemainingResult<(Vec<To
                 PRIORITY_DEAD_CODE,
                 format!("Unused function: {}", func.name),
             )
-            .with_location(func.file.display().to_string(), func.line as u32)
+            .with_location(func.file.display().to_string(), func.line)
             .with_severity("medium")
         })
         .collect();
